Moved non-Function content to scad.go. Added package description.
